+++ components/t_section.js
@@ -7,7 +7,7 @@ export default function T_section() {
         <div class="py-8 px-4 mx-auto max-w-screen-xl lg:py-16 lg:px-6 ">
           <div class="mx-auto max-w-screen-sm text-center mb-8 lg:mb-16">
             <h2 class="mb-4 text-4xl tracking-tight font-extrabold text-gray-900 dark:text-white">Meet Our Expert Team</h2>
-            <p class="font-light text-gray-500 lg:mb-16 sm:text-xl dark:text-gray-400">We are all passionate and impact-driven </p>
+            <p class="mb-3 font-normal text-dark-700 dark:text-dark-400">We are all passionate and impact-driven </p>
           </div>
           <div class="grid gap-8 mb-6 lg:mb-16 md:grid-cols-2">
             {team_list.map((team_list) => (
@@ -19,8 +19,8 @@ export default function T_section() {
                   <h3 class="text-xl font-bold tracking-tight text-gray-900 dark:text-white">
                     <a href="#">{team_list.name}</a>
                   </h3>
-                  <span class="text-gray-500 font-bold dark:text-gray-400">{team_list.role}</span>
-                  <p class="mt-3 mb-4 font-light italic text-gray-500 dark:text-gray-400">{team_list.experience}</p>
+                  <span class="text-black-500 font-bold dark:text-gray-400">{team_list.role}</span>
+                  <p class="mt-3 mb-4 font-light italic text-dark-500 dark:text-gray-400">{team_list.experience}</p>
                   <ul class="flex space-x-4 sm:mt-0">
                     <li>
                       <a href={team_list.website} class="text-gray-500 hover:text-gray-900 dark:hover:text-white">
