--- components/team-list.js
@@ -1,45 +1,0 @@
-const team_list = [
-     {
-       name: "Dr. Assetou F. SAMAKE",
-       image: "/images/assetou_foune_samake.jpeg",
-       experience: "Former Malian Minister of Higher Education, Biology Lecturer",
-       role:
-         "Co-founder, Board member, Advisor on Business Sustainability and external relations",
-       website: "https://fr.wikipedia.org/wiki/Ass%C3%A9tou_Foun%C3%A9_Samak%C3%A9_Migan",
-     },
-   
-     {
-       name: "Michael LEVENTHAL",
-       image: "/images/michael_leventhal.jpeg",
-       experience: "Experienced Senior & Product Manager, Founder and Board President of RobotsMali",
-       role:
-         "Co-founder, Board member, Advisor on Technical related topics/areas and international relations",
-       website: "https://www.linkedin.com/in/michael-leventhal-87239/",
-     },
-     {
-       name: "Massa TRAORE",
-       image: "/images/massa.jpeg",
-       experience: "Experienced Entrepreuner, Knight of the National Order",
-       role:
-         "Co-founder, Board member, Advisor on business and external relations",
-       website: "",
-     },
-     {
-       name: "Mamadou K. KEITA",
-       image: "/images/mk.jpg",
-       experience: "Computer Science, Robotics & AI, Co-founder of Totally Design Corp",
-       role:
-         "Founder, Board Member",
-       website: "https://mamadouk.me",
-     },
-     {
-       name: "Abdoulaye KEITA",
-       image: "/images/avatar.png",
-       experience: "Lieutnant  Malian Firefighter",
-       role:
-         "Advisor on security and robots reliablity on field",
-       website: "https://mamadouk.me",
-     },
-   ];
-   
-   export default team_list;
