+++ components/team_list.js
@@ -1,0 +1,20 @@
+const team_list = [
+     {
+       name: "Mamadou K. KEITA",
+       image: "/assets/mk.jpeg",
+       experience: "Computer Science, Robotics & AI, Co-founder of Totally Design Corp",
+       role:
+         "Founder, Board Member",
+       website: "https://mamadouk.me",
+     },
+     {
+       name: "Abdoulaye SAKO",
+       image: "/images/avatar.png",
+       experience: "Robotics & AI",
+       role:
+         "Mechanical engineer",
+       website: "/",
+     },
+   ];
+   
+   export default team_list;
